requirements.txt
python-dotenv
pandas
numpy
certifi
scikit-learn
pymongo>=4.3.3
dill ## uselful to pickle our file
pyaml
mlflow==2.9.2
dagshub
fastapi
uvicorn  ## both this will require for frontend and app.py
# -e . ## -e. refers to -e . setup.py  build package from setup.py
## we can put it at last when our entire packages is built

# useful when you try to deploy you package in cloud or pypy